Clean up stores
