+++ dsp/main.js
@@ -1,6 +1,7 @@
 import { Renderer, el } from '@elemaudio/core';
 import { RefMap } from './RefMap';
 import srvb from './srvb';
+import frequencyShift from './hilbertShift';
 
 
 // This project demonstrates writing a small FDN reverb effect in Elementary.
@@ -33,6 +34,16 @@ globalThis.__receiveStateChange__ = (serializedState) => {
   const state = JSON.parse(serializedState);
 
   if (shouldRender(prevState, state)) {
+
+    let fs = frequencyShift({
+      key: 'freqshift',
+      sampleRate: state.sampleRate,
+      shift: refs.getOrCreate('shift', 'const', { value: state.shift === 0 ? 0.0001 : state.shift }, []),
+    },
+    el.in({ channel: 0 }), el.in({ channel: 1 })
+    );
+
+
     let stats = core.render(...srvb({
       key: 'srvb',
       sampleRate: state.sampleRate,
@@ -40,17 +51,18 @@ globalThis.__receiveStateChange__ = (serializedState) => {
       decay: refs.getOrCreate('decay', 'const', { value: state.decay }, []),
       mod: refs.getOrCreate('mod', 'const', { value: state.mod }, []),
       mix: refs.getOrCreate('mix', 'const', { value: state.mix }, []),
-    }, el.in({ channel: 0 }), el.in({ channel: 1 })));
+    }, fs[0], fs[1]));
 
     console.log(stats);
+  
   } else {
     console.log('Updating refs');
     refs.update('size', { value: state.size });
     refs.update('decay', { value: state.decay });
     refs.update('mod', { value: state.mod });
     refs.update('mix', { value: state.mix });
+    refs.update('shift', { value: state.shift });
   }
-
   prevState = state;
 };
 
@@ -80,5 +92,5 @@ globalThis.__receiveHydrationData__ = (data) => {
 
 // Finally, an error callback which just logs back to native
 globalThis.__receiveError__ = (err) => {
-  console.log(`[Error: ${err.name}] ${err.message}`);
+  console.log(`[Elem Error: ${err.name}] ${err.message}`);
 };
