+++ public/manifest.json
@@ -7,7 +7,7 @@
     { "paramId": "size", "name": "Size", "min": 0.0, "max": 1.0, "defaultValue": 0.25 },
     { "paramId": "decay", "name": "Decay", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
     { "paramId": "mod", "name": "Mod", "min": 0.0, "max": 1.0, "defaultValue": 0.5 },
-    { "paramId": "mix", "name": "Mix", "min": 0.0, "max": 1.0, "defaultValue": 1.0 }
+    { "paramId": "mix", "name": "Mix", "min": 0.0, "max": 1.0, "defaultValue": 1.0 },
+    {"paramId": "shift", "name": "Shift", "min": 0.0, "max": 0.5, "defaultValue": 0.0 }
   ]
 }
-
